fix: 구조 수정
