+++ src/main/java/com/cagong/receiptpowerserver/domain/member/MemberRepository.java
@@ -2,5 +2,8 @@ package com.cagong.receiptpowerserver.domain.member;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.Optional;
+
 public interface MemberRepository extends JpaRepository<Member, Long> {
+    Optional<Member> findByUsername(String username);
 }
