+++ src/main/java/com/cagong/receiptpowerserver/domain/member/dto/MemberLoginDto.java
@@ -1,4 +1,4 @@
-package com.cagong.receiptpowerserver.member.dto;
+package com.cagong.receiptpowerserver.domain.member.dto;
 
 import lombok.Getter;
 import lombok.Setter;
