--- src/main/java/com/cagong/receiptpowerserver/global/config/SecurityConfig.java
@@ -1,4 +1,0 @@
-package com.cagong.receiptpowerserver.global.config;
-
-public class SecurityConfig {
-}
