--- src/main/java/com/cagong/receiptpowerserver/member/Member.java
@@ -1,24 +1,0 @@
-package com.cagong.receiptpowerserver.member;
-
-import com.cagong.receiptpowerserver.common.BaseEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Member extends BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, unique = true, length = 20)
-    private String username;
-
-    @Column(nullable = false)
-    private String password;
-
-    // 추가 필드(이메일, 닉네임 등)는 필요 시 여기에 확장
-}
