--- src/main/java/com/cagong/receiptpowerserver/member/MemberController.java
@@ -1,36 +1,0 @@
-package com.cagong.receiptpowerserver.member;
-
-import com.cagong.receiptpowerserver.member.dto.MemberRequestDto;
-import com.cagong.receiptpowerserver.member.dto.MemberLoginDto;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.*;
-import org.springframework.web.bind.annotation.*;
-import org.springframework.security.authentication.BadCredentialsException;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-
-@RestController
-@RequestMapping("/api/member")
-@RequiredArgsConstructor
-public class MemberController {
-    private final MemberService memberService;
-
-    @PostMapping("/signup")
-    public ResponseEntity<String> signup(@RequestBody MemberRequestDto dto) {
-        try {
-            memberService.signup(dto);
-            return ResponseEntity.status(HttpStatus.CREATED).body("회원가입 성공");
-        } catch (IllegalArgumentException e) {
-            return ResponseEntity.status(HttpStatus.CONFLICT).body(e.getMessage());
-        }
-    }
-
-    @PostMapping("/login")
-    public ResponseEntity<String> login(@RequestBody MemberLoginDto dto) {
-        try {
-            String jwt = memberService.login(dto);
-            return ResponseEntity.ok(jwt);
-        } catch (UsernameNotFoundException | BadCredentialsException e) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(e.getMessage());
-        }
-    }
-}
