--- src/main/java/com/cagong/receiptpowerserver/member/MemberRepository.java
@@ -1,9 +1,0 @@
-package com.cagong.receiptpowerserver.member;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-    Optional<Member> findByUsername(String username);
-}
